+++ sdk/provisioning/Azure.Provisioning/src/Expressions/AST/NullLiteralExpression.cs
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Azure.Provisioning.Expressions;
+
+public class NullLiteralExpression() : LiteralExpression()
+{
+    internal override BicepWriter Write(BicepWriter writer) => writer.Append("null");
+}
